Fix remaining 13 errors: execution_loop.rs, work_integration.rs, acceptance.rs, golden_paths.rs, review.rs, file_control.rs, failure.rs

+++ src/harness/golden_paths.rs
@@ -512,7 +512,7 @@ impl GoldenPathRegistry {
             }
 
             // Calculate match score
-            let mut score: f32 = 0.0;
+            let mut score: f64 = 0.0;
             let mut reasons = Vec::new();
 
             // Check keywords in context
